Implemented DLL Header parsing for windows
This allows the REGISTER_FIXTURE to be used immediately after the class
declaration rather than in DLL main. The old REGISTER_FIXTURE has been
renamed to REGISTER_FIXTURE_MAIN to indicate that it must be used in
dllMain. This is for OS that do not yet have DLL header parsing